+++ src/spellbot/cogs/__init__.py
@@ -13,6 +13,7 @@ from .block_cog import BlockCog
 from .events_cog import EventsCog
 from .leave_cog import LeaveGameCog
 from .lfg_cog import LookingForGameCog
+from .mod_cog import ModCog
 from .owner_cog import OwnerCog
 from .score_cog import ScoreCog
 from .sync_cog import SyncCog
@@ -30,6 +31,7 @@ __all__ = [
     "EventsCog",
     "LeaveGameCog",
     "LookingForGameCog",
+    "ModCog",
     "OwnerCog",
     "ScoreCog",
     "SyncCog",